test(#1293): delete obsolete Issue1287 Groq tests, rewrite Issue864 via adapter API, add dead-symbol removal guard (RED)
